+++ frontend/src/features/events/components/calendar/EventsCalendar.jsx
@@ -6,7 +6,7 @@ import {
 } from 'react-icons/hi';
 import { useCalendar } from './useCalendar';
 import { STATUS_METADATA } from '../../utils/constants';
-import { withAlpha } from '../../utils/formats';
+import { withAlpha } from '../../utils/format';
 
 const viewOptions = [
   { value: 'month', label: 'Month' },
